Implemented max flow algorithm, and adjusted test

+++ week2_part1/testing/unittests/predefined_week2_part1_unittests.cpp
@@ -18,8 +18,8 @@ namespace alg = dte3603::week2::algorithms;
 // DAGOneF
 TEST_F(DGOneF, getMaxFlowTest)
 {
-  int flow = alg::getMaxFlow(gold->graph(), gold->s(), gold->t());
-  int gold_flow = gold->flowFromStoTGold();
+  double flow      = alg::getMaxFlow(gold->graph(), gold->s(), gold->t());
+  double gold_flow = gold->flowFromStoTGold();
 
   EXPECT_EQ(flow, gold_flow);
 }
